(Need help) Started work on search feature
So I have come to the point where a search result has been selected. Now
the only thing that's left is to acually make the choosen file selected
in the editor. I don't really know how to do this as it's hard to
understand the editor tree system. To select an entry this should be
used: treeTable.setRowSelectionInterval(i, i); But somehow it needs to
loop thought the whole tree to get the index of the right file